# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/insights:
    post:
      operationId: streamAnalyticsInsights
      tags: [analytics]
      summary: Stream AI-generated coaching insights based on current analytics data (SSE)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyticsInsightsInput"
      responses:
        "200":
          description: SSE stream of insight text chunks
          content:
            text/event-stream: {}
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk session JWT. The browser sends this as a cookie managed by the Clerk SDK; mobile clients send it in the Authorization header as `Bearer <token>`.
  schemas:
...
    AnalyticsInsightsInput:
      type: object
      properties:
        focus:
          type: string
          nullable: true
          description: Optional focus area for the insights (e.g. "conversion", "sequences")
...